// DocsFirst/src/server/types.ts
export const MessageType = {
	Unknown: -1,
	Sync: 0,
	Awareness: 1,
	Auth: 2,
	QueryAwareness: 3,
	SyncReply: 4, // same as Sync, but won't trigger another 'SyncStep1'
	Stateless: 5,
	BroadcastStateless: 6,
	CLOSE: 7,
	SyncStatus: 8,
} as const;

export type MessageType = typeof MessageType[keyof typeof MessageType];